:green_heart: :recycle: Extract type_simple_test.go and type_string_test.go
