+++ parser/type_test.go
@@ -250,107 +250,3 @@ func TestNamedType(t *testing.T) {
 	run([]rune(`CURRENCY`), &ast.RealType{Name: ast.Ident("CURRENCY")})
 	run([]rune(`COMP`), &ast.RealType{Name: ast.Ident("COMP")})
 }
-
-func TestEnumeratedType(t *testing.T) {
-	run := func(name string, text []rune, expected ast.Type) {
-		t.Run(name, func(t *testing.T) {
-			parser := NewParser(&text)
-			parser.NextToken()
-			res, err := parser.ParseType()
-			if assert.NoError(t, err) {
-				assert.Equal(t, expected, res)
-			}
-		})
-	}
-
-	run(
-		"card",
-		[]rune(`(Club, Diamond, Heart, Spade)`),
-		ast.EnumeratedType{
-			{Ident: ast.Ident("Club")},
-			{Ident: ast.Ident("Diamond")},
-			{Ident: ast.Ident("Heart")},
-			{Ident: ast.Ident("Spade")},
-		},
-	)
-
-	run(
-		"Enumerated types with explicitly assigned ordinality",
-		[]rune(`(Small = 5, Medium = 10, Large = Small + Medium)`),
-		ast.EnumeratedType{
-			{Ident: ast.Ident("Small"), ConstExpr: ast.NewConstExpr(ast.NewNumber("5"))},
-			{Ident: ast.Ident("Medium"), ConstExpr: ast.NewConstExpr(ast.NewNumber("10"))},
-			{Ident: ast.Ident("Large"), ConstExpr: ast.NewConstExpr(
-				&ast.SimpleExpression{
-					Term: *ast.NewTerm("Small"),
-					AddOpTerms: []*ast.AddOpTerm{
-						{AddOp: "+", Term: *ast.NewTerm("Medium")},
-					},
-				},
-			)},
-		},
-	)
-
-}
-
-func TestSubrangeType(t *testing.T) {
-	run := func(name string, text []rune, expected ast.Type) {
-		t.Run(name, func(t *testing.T) {
-			parser := NewParser(&text)
-			parser.NextToken()
-			res, err := parser.ParseType()
-			if assert.NoError(t, err) {
-				assert.Equal(t, expected, res)
-			}
-		})
-	}
-
-	run(
-		"subrange type of enumerated type",
-		[]rune(`Green..White`),
-		&ast.SubrangeType{Low: *ast.NewConstExpr("Green"), High: *ast.NewConstExpr("White")},
-	)
-	run(
-		"subrange type of number",
-		[]rune(`-128..127`),
-		&ast.SubrangeType{Low: *ast.NewConstExpr(ast.NewNumber("-128")), High: *ast.NewConstExpr(ast.NewNumber("127"))},
-	)
-	run(
-		"subrange type of character",
-		[]rune(`'A'..'Z'`),
-		&ast.SubrangeType{Low: *ast.NewConstExpr(ast.NewString("'A'")), High: *ast.NewConstExpr(ast.NewString("'Z'"))},
-	)
-}
-
-func TestStringType(t *testing.T) {
-	run := func(name string, text []rune, expected ast.Type) {
-		t.Run(name, func(t *testing.T) {
-			parser := NewParser(&text)
-			parser.NextToken()
-			res, err := parser.ParseType()
-			if assert.NoError(t, err) {
-				assert.Equal(t, expected, res)
-			}
-		})
-	}
-
-	run("String", []rune(`STRING`), &ast.StringType{Name: "STRING"})
-	run("ANSI String", []rune(`ANSISTRING`), &ast.StringType{Name: "ANSISTRING"})
-	run("Wide String", []rune(`WIDESTRING`), &ast.StringType{Name: "WIDESTRING"})
-	run("Short String", []rune(`STRING[100]`), &ast.StringType{Name: "STRING", Length: ast.NewConstExpr(ast.NewNumber("100"))})
-	run(
-		"Short String",
-		[]rune(`STRING[ALen + BLen]`),
-		&ast.StringType{
-			Name: "STRING",
-			Length: ast.NewConstExpr(
-				&ast.SimpleExpression{
-					Term: *ast.NewTerm("ALen"),
-					AddOpTerms: []*ast.AddOpTerm{
-						{AddOp: "+", Term: *ast.NewTerm("BLen")},
-					},
-				},
-			),
-		},
-	)
-}
